+++ source/Bot/Models/Booru/EmbedPost.cs
@@ -1,4 +1,4 @@
-namespace Bot.Models
+namespace Bot.Models.Booru
 {
