+++ source/Bot/Models/Casino/BlackJack/BlackJackPlayer.cs
@@ -1,3 +1,4 @@
+using Bot.Models.Eileen;
 using Discord;
 
 namespace Bot.Models.Casino.BlackJack
