+++ source/Bot/Models/Casino/CasinoPlayer.cs
@@ -1,3 +1,4 @@
+using Bot.Models.Eileen;
 using Discord;
 
 namespace Bot.Models.Casino
