+++ source/Bot/Models/Eileen/ServerConfigurationData.cs
@@ -1,4 +1,4 @@
-namespace Bot.Models
+namespace Bot.Models.Eileen
 {
