+++ source/Bot/Profiles/GelbooruProfile.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using Bot.Models.Booru;
 
 namespace Bot.Profiles
 {
@@ -8,7 +9,7 @@ namespace Bot.Profiles
 
         public GelbooruProfile()
         {
-            CreateMap<Models.Gelbooru.Post, Models.EmbedPost>()
+            CreateMap<Models.Gelbooru.Post, EmbedPost>()
                 .ForMember(dest => dest.ArtistName, opt => opt.MapFrom(src => string.IsNullOrWhiteSpace(src.Owner) ? "N/A" : src.Owner))
                 .ForMember(dest => dest.ImageUrl, opt => opt.MapFrom(src => src.FileUrl))
                 .ForMember(dest => dest.PageUrl, opt => opt.MapFrom(src => $"https://gelbooru.com/index.php?page=post&s=view&id={src.Id}"));
