+++ source/Bot/Profiles/SafebooruProfile.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using Bot.Models.Booru;
 
 namespace Bot.Profiles
 {
@@ -8,7 +9,7 @@ namespace Bot.Profiles
 
         public SafebooruProfile()
         {
-            CreateMap<Models.Safebooru.Post, Models.EmbedPost>()
+            CreateMap<Models.Safebooru.Post, EmbedPost>()
                 .ForMember(dest => dest.ArtistName, opt => opt.MapFrom(src => string.IsNullOrWhiteSpace(src.Owner) ? "N/A" : src.Owner))
                 .ForMember(dest => dest.ImageUrl, opt => opt.MapFrom(src => $"https://safebooru.org/images/{src.Directory}/{src.Image}"))
                 .ForMember(dest => dest.PageUrl, opt => opt.MapFrom(src => $"https://safebooru.org/index.php?page=post&s=view&id={src.Id}"));
