+++ source/Bot/Profiles/e621Profile.cs
@@ -1,4 +1,5 @@
 using AutoMapper;
+using Bot.Models.Booru;
 using System.Linq;
 
 namespace Bot.Profiles
@@ -9,7 +10,7 @@ namespace Bot.Profiles
 
         public e621Profile()
         {
-            CreateMap<Models.e621.Post, Models.EmbedPost>()
+            CreateMap<Models.e621.Post, EmbedPost>()
                 .ForMember(dest => dest.ArtistName, opt => opt.MapFrom(src => src.Tags.Artist.Any() ? string.Join(",", src.Tags.Artist) : "N/A"))
                 .ForMember(dest => dest.ImageUrl, opt => opt.MapFrom(src => src.File.Url))
                 .ForMember(dest => dest.PageUrl, opt => opt.MapFrom(src => src.GetPostUrl()));
