+++ source/Bot/Program.cs
@@ -1,5 +1,5 @@
 using AutoMapper;
-using Bot.Models;
+using Bot.Models.Eileen;
 using Bot.Services;
 using Bot.Services.RavenDB;
 using Discord;
