+++ source/Bot/Services/CredentialsService.cs
@@ -1,3 +1,4 @@
+using Bot.Models.Booru;
 using Bot.Services.RavenDB;
 using System.Collections.Generic;
 using System.Linq;
@@ -8,7 +9,7 @@ namespace Bot.Services
     public sealed class CredentialsService : IEileenService
     {
 
-        public IReadOnlyList<Models.CredentialsEntry> Credentials { get; private set; }
+        public IReadOnlyList<CredentialsEntry> Credentials { get; private set; }
 
 
         public CredentialsService(RavenDatabaseService rdbs)
