+++ source/Bot/Services/ServerConfigurationService.cs
@@ -1,4 +1,4 @@
-using Bot.Models;
+using Bot.Models.Eileen;
 using Bot.Services.RavenDB;
 using Discord;
 using Discord.Commands;
